atom-bionic: Install Composer
AtoM 2.6.0+ will use Composer to manage some third-party PHP libraries.

- Install php-cli before installing Composer:

PHP-cli is required to run the Composer installer, but because PHP is
installed in the ansible-atom it comes too late for the composer role.
Installing php-cli in the pre-tasks and again in ansible-atom isn't
ideal, but it avoids having to move PHP installation entirely from
ansible-atom to this playbook.

- Install Composer with geerlingguy/ansible-role-composer.