Run AM playbook test workflow in Ubuntu 20.04
The workflow started failing after the 20240915.1.0 release of the
Ubuntu 22.04 runner image because VirtualBox cannot be downgraded to
version 7.0.14 anymore since it depends on older libraries to compile
its drivers.

Upgrading to a newer version of VirtualBox requires updates to the
network setup of the VM in the Vagrantfile of the playbook.